+++ EssentialApp/EssentialAppTests/FeedImageDataLoaderCacheDecoratorTests.swift
@@ -73,8 +73,8 @@ final class FeedImageDataLoaderCacheDecoratorTests: XCTestCase {
     private func makeSUT(
         file: StaticString = #file,
         line: UInt = #line
-    ) -> (sut: FeedImageDataLoader, loader: LoaderSpy) {
-        let loader = LoaderSpy()
+    ) -> (sut: FeedImageDataLoader, loader: FeedImageDataLoaderSpy) {
+        let loader = FeedImageDataLoaderSpy()
         let sut = FeedImageDataLoaderCacheDecorator(decoratee: loader)
 
         trackMemoryLeaks(loader)
@@ -109,38 +109,5 @@ final class FeedImageDataLoaderCacheDecoratorTests: XCTestCase {
         action()
 
         wait(for: [exp], timeout: 1.0)
-    }
-
-    private class LoaderSpy: FeedImageDataLoader {
-        private var messages = [(url: URL, completion: (FeedImageDataLoader.Result) -> Void)]()
-
-        var loadedURLs: [URL] {
-            messages.map{ $0.url }
-        }
-
-        private(set) var cancelledURLs = [URL]()
-
-        private struct Task: FeedImageDataLoaderTask {
-            let callback: () -> (Void)
-            func cancel() {
-                callback()
-            }
-        }
-
-        func loadImageData(from url: URL, completion: @escaping (FeedImageDataLoader.Result) -> Void) -> FeedImageDataLoaderTask {
-            messages.append((url, completion))
-
-            return Task { [weak self] in
-                self?.cancelledURLs.append(url)
-            }
-        }
-
-        func complete(with error: NSError, and index: Int = 0) {
-            messages[index].completion(.failure(error))
-        }
-
-        func complete(with expectedData: Data, and index: Int = 0) {
-            messages[index].completion(.success(expectedData))
-        }
     }
 }
